Refactor streams query semantic analysis to a new class
